requirements.txt
asgiref==3.11.1
Django==5.2.12
django-cors-headers==4.9.0
django-filter==25.2
django-phonenumber-field==8.4.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
phonenumbers==9.0.26
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
